+++ app/client/src/pages/CartPage.js
@@ -1,8 +1,9 @@
 import React, { useEffect, useState } from 'react';
 import Header from '../components/Header/Header';
 import CartList from '../components/CartList/CartList';
-import Checkout from '../components/Checkout';
-import './Page.css'
+import Checkout from '../components/Checkout/Checkout';
+import Page from '../components/Page/Page'
+import EmptyCart from '../components/EmptyCart';
 
 const nearestHundredth = (val) => parseFloat(Math.round(val * 100)) / 100
 
@@ -62,13 +63,20 @@ export default function CartPage({activeUser}) {
         fetchCart();
     }, []);
     console.log(cart);
-    return (
-        <div className='page-body'>
-            <Header loggedIn={false}/>
-            <section className='main-content'>
+    const mainBody = () => {
+        if(cart.length === 0){
+            return <EmptyCart/>
+        }
+        else{
+            return <>
                 <CartList items={cart} handleUpdate={updateCart} setTotal={setTotal} total={total}/>
                 <Checkout total={total} handleOrder={order}/>
-            </section>
-        </div>
+            </>
+        }
+    }
+    return (
+        <Page activeUser={activeUser}>
+            {mainBody()}
+        </Page>
     );
 }